# doc/swagger/storage.yaml
swagger: '2.0'
info:
  description: This service provides access to Storage API
  version: 1.0.0
  title: Storage DataService API
host: localhost
basePath: /storage
schemes:
  - http
  - https
...
paths:
  /:
    post:
      summary: Set all key/value pairs in Storage
      description: Set all key/value pairs in Storage using object as source.
      operationId: setAll
      tags:
        - Storage
      consumes:
        - application/json
      parameters:
        - $ref: '#/parameters/storageType'
        - $ref: '#/parameters/dict'
      responses:
        204:
          description: Success
        500:
          description: Error
          schema:
            $ref: '#/definitions/Error'
    get:
      summary: Get all key/value pairs from Storage
      description: Get all key/value pairs as object from Storage
      operationId: getAll
      tags:
        - Storage
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/storageType'
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/Dict'
        500:
          description: Error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Delete all key/value pairs from Storage
      description: Clear Storage
      operationId: deleteAll
      tags:
        - Storage
      parameters:
        - $ref: '#/parameters/storageType'
      responses:
        204:
          description: Success
        500:
          description: Error
          schema:
            $ref: '#/definitions/Error'
...
parameters:
  storageType:
    in: query
    name: storageType
    required: false
    type: string
    enum: [local, cluster, ha]
    description: Storage location type
...
  dict:
    in: body
    required: true
    name: dict
    schema:
      $ref: '#/definitions/Dict'
definitions:
  AnyValue: {}
  Dict:
    type: object
    additionalProperties:
      $ref: '#/definitions/AnyValue'
    example:
      key1: value1
      key2: value2
  Error:
    type: object
    properties:
      err:
        type: string
    example:
      err: "An error occurred"